+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Released]: v2.0.9
+- Add logs for Verifier 403 response
+- Update Gradle Sonar version
+
 ## [Released]: v2.0.8
 - Update version entity validation to support three-level versioning